requirements.txt
Django>=3.9,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2~=2.9.5
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
djangorestframework-simplejwt>=4.9.0, <5.2.1
django-cors-headers>=3.7.0,<4.1.0
uwsgi>=2.0.19<2.1
